Crate bklyn

Bklyn is a query interface for the kubernetes cluster metric service heapster.

Structs

Aggregate

Aggregates

Cluster
Metrics associated with a kubernetes cluster

FreeContainer
Metrics associated with a container not bound to a specific pod

Heapster
Central interface for communicating kubernetes heapster service

MetricCollection

MetricOptions
query options for fetching metric values

MetricOptionsBuilder

Metrics
metric interface

Namespace
Metrics within a cluster namespace

NamespacePod
Metrics associated with a pod within a given namespace

NamespacePodContainer
Metrics associated with a container, within a pod, within a namespace

Node
A node is essentially a host within a cluster

Stats

Summary

Value

Enums

Credentials
Credentials used for authenticating with kubernetes cluster

Error
enumerated types of client errors

Type Definitions

Result
Result type for bklyn query operations